# plans/openapi.json.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrailTag API",
    "description": "YouTube 旅遊影片地圖化 API 服務",
    "version": "0.1.0"
  },
  "paths": {
    "/api/videos/analyze": {
      "post": {
        "summary": "提交影片分析任務",
        "description": "提交新的 YouTube 影片分析請求，支援快取查詢與非同步任務分派。回傳 job_id 供進度查詢。",
        "operationId": "analyze_video_api_videos_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "summary": "查詢任務狀態",
        "description": "根據 job_id 查詢分析任務的進度、階段與錯誤資訊。",
        "operationId": "get_job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任務 ID",
              "title": "Job Id"
            },
            "description": "任務 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/{video_id}/locations": {
      "get": {
        "summary": "查詢影片地點視覺化資料",
        "description": "根據 YouTube video_id 查詢分析後的地點視覺化資料（地圖路線等）。",
        "operationId": "get_video_locations_api_videos__video_id__locations_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "YouTube 影片 ID",
              "title": "Video Id"
            },
            "description": "YouTube 影片 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapVisualization"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/stream": {
      "get": {
        "summary": "Stream Job Events",
        "description": "以 SSE 格式推送指定任務的即時進度、狀態與錯誤事件。\n前端可透過此端點即時監控任務進度。",
        "operationId": "stream_job_events_api_jobs__job_id__stream_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任務 ID",
              "title": "Job Id"
            },
            "description": "任務 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "基本健康檢查端點",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Check Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url",
            "description": "YouTube 影片 URL"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AnalyzeRequest",
        "description": "分析請求模型"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Phase"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "video_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse",
        "description": "任務回應模型"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "done",
          "failed",
          "canceled"
        ],
        "title": "JobStatus",
        "description": "任務狀態列舉"
      },
      "MapVisualization": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "routes": {
            "items": {
              "$ref": "#/components/schemas/RouteItem"
            },
            "type": "array",
            "title": "Routes"
          }
        },
        "type": "object",
        "required": [
          "video_id",
          "routes"
        ],
        "title": "MapVisualization",
        "description": "地圖視覺化資料，包含影片ID與多個路線項目。"
      },
      "Phase": {
        "type": "string",
        "enum": [
          "metadata",
          "compression",
          "summary",
          "geocode"
        ],
        "title": "Phase",
        "description": "處理階段列舉"
      },
      "RouteItem": {
        "properties": {
          "location": {
            "type": "string",
            "title": "Location"
          },
          "coordinates": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinates"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "timecode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timecode"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "marker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marker"
          }
        },
        "type": "object",
        "required": [
          "location"
        ],
        "title": "RouteItem",
        "description": "路線項目，描述地圖上一個地點與相關資訊。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}